Add files via upload
Tumor detection project, the dataset of this project is taken from kaggle , it include total 225 images for training ,20 for validation, & 17 for test. It's hard to work on a project containing a small no. of images for training but i try my best and used 4 techniques to increase my model accuracy including Data Augmentation , feature extraction from a pretrained model and apply those features as numpy array in the new model , use a pretrained vgg16 model and attached a densely connected classifier as well as fine tuned the model parameters.